apps/voice-agent/requirements.txt
# LiveKit agents framework
livekit>=0.12.0
livekit-agents>=0.8.0
livekit-plugins-silero>=0.6.0

# Speech-to-text (faster-whisper uses CTranslate2)
faster-whisper>=1.0.0

# Text-to-speech (Python bindings, CLI is installed separately)
piper-tts>=1.2.0

# LLM client (OpenAI-compatible for Nebius)
openai>=1.0.0

# Environment and utilities
python-dotenv>=1.0.0

# Audio processing
numpy>=1.24.0
soundfile>=0.12.0

# GPU support (optional - for faster-whisper)
# Uncomment if using GPU:
# torch>=2.0.0
# ctranslate2>=4.0.0
